Move ServiceNetworkSender to instance::network_sender
This is part of a series of commits to move the existing APIs for
services into a new modules to make room for the new threadless design.
