+++ libsplinter/src/service/instance/mod.rs
@@ -14,8 +14,10 @@
 
 mod message_context;
 mod network_registry;
+mod network_sender;
 mod service_instance;
 
 pub use message_context::ServiceMessageContext;
 pub use network_registry::ServiceNetworkRegistry;
+pub use network_sender::ServiceNetworkSender;
 pub use service_instance::ServiceInstance;
