+++ libsplinter/src/service/processor/mod.rs
@@ -657,7 +657,7 @@ pub mod tests {
     use crate::service::error::{
         ServiceDestroyError, ServiceError, ServiceStartError, ServiceStopError,
     };
-    use crate::service::{instance::ServiceNetworkRegistry, ServiceNetworkSender};
+    use crate::service::instance::{ServiceNetworkRegistry, ServiceNetworkSender};
     use crate::transport::inproc::InprocTransport;
     use crate::transport::matrix::ConnectionMatrixSender;
     use crate::transport::Transport;
